backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.0
langchain-google-genai==2.0.0
langgraph==0.2.28
langchain-community==0.3.0
pypdf2==3.0.1
pdfplumber==0.11.4
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2
arxiv==2.1.3
requests==2.32.3
pydantic==2.9.2
